feat: added all the interfaces
